Add env hooks to influence building and linking with expat.
We build our own expat libraries so they are not installed in the
standard location. Therefore, we need to customize the build so it links
against the correct libraries. Normally, you could just pass EXPATLIBPATH
and EXPATINCPATH parameters to Makefile.PL. But we added these environment
variable hooks so that you can set those parameters whilst using cpanm.